# api-spec.yaml
openapi: 3.0.3
info:
  title: 40Labs Healthcare Platform API - Device Activation V2
  version: 2.0.0
paths:
  /api/v1/activation/otp/request:
    post:
      summary: Request OTP for device activation
      description: Sends a one-time password (OTP) via SMS to the owner phone number.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phone
              properties:
                phone:
                  type: string
                  example: "+255712345678"
      responses:
        '200':
          description: OTP sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "otp_sent"
                  expires_in_secs:
                    type: integer
                    example: 300
        '400':
          description: Invalid phone number
        '429':
          description: Rate limit exceeded

  /api/v1/activation/otp/verify:
    post:
      summary: Verify OTP code
      description: Verifies the SMS OTP code and returns a single-use verification token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phone
                - code
              properties:
                phone:
                  type: string
                code:
                  type: string
      responses:
        '200':
          description: OTP verified
          content:
            application/json:
              schema:
                type: object
                properties:
                  verification_token:
                    type: string
        '400':
          description: Invalid or expired OTP code

  /api/v1/activation/register:
    post:
      summary: Register primary device and business activation
      description: Registers business details, owner account, and terms acceptance. Returns cryptographically signed ActivationGrant, business_id, workspace_id, and branch_id.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - verification_token
                - business_name
                - owner_first_name
                - owner_last_name
                - terms_accepted
                - terms_version
                - terms_locale
                - terms_text_sha256
              properties:
                verification_token:
                  type: string
                business_name:
                  type: string
                owner_first_name:
                  type: string
                owner_last_name:
                  type: string
                owner_email:
                  type: string
                  nullable: true
                terms_accepted:
                  type: boolean
                terms_version:
                  type: string
                terms_locale:
                  type: string
                terms_text_sha256:
                  type: string
      responses:
        '200':
          description: Device and business activated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  activation_grant:
                    type: string
                    description: Cryptographically signed grant for primary device
                  business_id:
                    type: string
                  workspace_id:
                    type: string
                  branch_id:
                    type: string
        '400':
          description: Invalid request or verification token
        '409':
          description: Business or phone already registered